general stats function as a module file
